+++ src/cljs/lambdacd/console_output_processor.cljs
@@ -1,5 +1,6 @@
 (ns lambdacd.console-output-processor
-  (:require [clojure.string :as s]))
+  (:require [clojure.string :as s]
+            [ansiparse]))
 
 (defn- process-carriage-returns [line]
   (->> (clojure.string/split line "\r")
@@ -17,7 +18,33 @@
           (= "\b" x) (recur xs (rest result))
           :else (recur xs (conj result x)))))))
 
-(defn process-ascii-escape-characters [s]
+(defn clean-up-text [s]
   (->> (s/split-lines s)
        (map process-carriage-returns)
-       (map process-backspaces)))
+       (map process-backspaces)
+       (s/join "\n")))
+
+(defn- de-ansify [s]
+  (if (= "" s)
+    [{:text s}]
+    (js->clj (js/ansiparse s) :keywordize-keys true)))
+
+(defn split-fragments-on-newline [fragment]
+  (cond
+    (= "\n" (:text fragment)) [:newline]
+    (= "" (:text fragment)) [fragment]
+    :else (->> (s/split-lines (:text fragment))
+               (map #(assoc fragment :text %))
+               (interpose :newline))))
+
+(defn- partition-by-newline [c]
+  (let [by-k #(not= :newline %)]
+    (->> (partition-by by-k c)
+         (filter #(not= [:newline] %)))))
+
+
+(defn process-ascii-escape-characters [s]
+  (->> (clean-up-text s)
+       (de-ansify)
+       (mapcat split-fragments-on-newline)
+       (partition-by-newline)))